[FIRMWARE] new joypad mapping functionality (which exposes up to 12 hardware buttons and can map from/to directions)

+++ usb/hid.h
@@ -50,6 +50,8 @@
 #define JOYSTICK_AXIS_TRIGGER_MIN   64
 #define JOYSTICK_AXIS_TRIGGER_MAX   192
 
+
+
 typedef struct {
   ep_t ep;    // interrupt endpoint info structure
 
@@ -64,8 +66,8 @@ typedef struct {
   
   // additional info extracted from the report descriptor
   // (currently only used for joysticks) 
-  uint8_t jmap;           // last reported joystick state
-  uint8_t jindex;         // joystick index
+  uint16_t jmap;           // last reported joystick state
+  uint16_t jindex;         // joystick index
   hid_report_t conf;
 
   uint8_t interval;
